+++ map/lib/di/map_module_factory.dart
@@ -1,4 +1,5 @@
 import 'package:core/di/core_module_factory.dart';
+import 'package:map/repository/camera_position_repository.dart';
 import 'package:map/usecase/get_map_provider_use_case.dart';
 import 'package:settings/di/settings_module_factory.dart';
 
@@ -7,5 +8,9 @@ class MapModuleFactory {
     return DefaultGetMapProviderUseCase(
         SettingsModuleFactory.createMapProviderRepository(),
         CoreModuleFactory.createConfigRepository());
+  }
+
+  static CameraPositionRepository createCameraPositionRepository() {
+    return LocalCameraPositionRepository();
   }
 }